# src/generatedSpringRestApi/java/com/github/tuczv/addressBook/api/EventsV1Api.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.github.tuczv.addressBook.api;

import com.github.tuczv.addressBook.api.dto.CalendarEventApiDto;
import com.github.tuczv.addressBook.api.dto.ErrorApiDto;
import com.github.tuczv.addressBook.api.dto.MailApiDto;
import java.util.UUID;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Tag(name = "EventsV1", description = "the EventsV1 API")
@RequestMapping("${openapi.addressbook.base-path:}")
public interface EventsV1Api {

    /**
     * POST /v1/events : Post a event.
     *
     * @return OK (status code 200)
     *         or error (status code 400)
     */
    @Operation(
        operationId = "createEvent",
        summary = "Post a event.",
        tags = { "events-v1" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CalendarEventApiDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorApiDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/v1/events",
        produces = { "application/json" }
    )
    ResponseEntity<CalendarEventApiDto> createEvent(
        
    );


    /**
     * DELETE /v1/emails/{emailId} : Get a email by Id.
     *
     * @param emailId Email id (required)
     * @return OK (status code 200)
     *         or error (status code 400)
     */
    @Operation(
        operationId = "findEmailById",
        summary = "Get a email by Id.",
        tags = { "events-v1" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = MailApiDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorApiDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/v1/emails/{emailId}",
        produces = { "application/json" }
    )
    ResponseEntity<MailApiDto> findEmailById(
        @Parameter(name = "emailId", description = "Email id", required = true) @PathVariable("emailId") UUID emailId
    );


    /**
     * GET /v1/events/{eventId} : Get a event by Id.
     *
     * @param eventId Event id (required)
     * @return OK (status code 200)
     *         or error (status code 400)
     */
    @Operation(
        operationId = "findEventById",
        summary = "Get a event by Id.",
        tags = { "events-v1" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CalendarEventApiDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorApiDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/v1/events/{eventId}",
        produces = { "application/json" }
    )
    ResponseEntity<CalendarEventApiDto> findEventById(
        @Parameter(name = "eventId", description = "Event id", required = true) @PathVariable("eventId") UUID eventId
    );


    /**
     * GET /v1/events : Get a calendar events. Results are pageable with seek method.
     *
     * @return OK (status code 200)
     *         or error (status code 400)
     */
    @Operation(
        operationId = "findEvents",
        summary = "Get a calendar events. Results are pageable with seek method.",
        tags = { "events-v1" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CalendarEventApiDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorApiDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/v1/events",
        produces = { "application/json" }
    )
    ResponseEntity<List<CalendarEventApiDto>> findEvents(
        
    );


    /**
     * PUT /v1/events/{eventId} : Update a event.
     *
     * @param eventId Event id (required)
     * @param calendarEventApiDto  (required)
     * @return OK (status code 200)
     *         or error (status code 400)
     */
    @Operation(
        operationId = "updateEvent",
        summary = "Update a event.",
        tags = { "events-v1" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CalendarEventApiDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorApiDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/v1/events/{eventId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<CalendarEventApiDto> updateEvent(
        @Parameter(name = "eventId", description = "Event id", required = true) @PathVariable("eventId") UUID eventId,
        @Parameter(name = "CalendarEventApiDto", description = "", required = true) @Valid @RequestBody CalendarEventApiDto calendarEventApiDto
    );

}