# src/generatedSpringRestApi/java/com/github/tuczv/addressBook/api/GroupsV1Api.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.github.tuczv.addressBook.api;

import com.github.tuczv.addressBook.api.dto.ErrorApiDto;
import com.github.tuczv.addressBook.api.dto.GroupApiDto;
import java.util.UUID;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated
@Tag(name = "GroupsV1", description = "the GroupsV1 API")
@RequestMapping("${openapi.addressbook.base-path:}")
public interface GroupsV1Api {

    /**
     * POST /v1/groups : Create a group.
     *
     * @param groupApiDto  (required)
     * @return OK (status code 200)
     *         or error (status code 400)
     */
    @Operation(
        operationId = "createGroup",
        summary = "Create a group.",
        tags = { "groups-v1" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GroupApiDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorApiDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/v1/groups",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<GroupApiDto> createGroup(
        @Parameter(name = "GroupApiDto", description = "", required = true) @Valid @RequestBody GroupApiDto groupApiDto
    );


    /**
     * GET /v1/groups/{groupId} : Get a group by Id.
     *
     * @param groupId Group id (required)
     * @return OK (status code 200)
     *         or error (status code 400)
     */
    @Operation(
        operationId = "findGroup",
        summary = "Get a group by Id.",
        tags = { "groups-v1" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GroupApiDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorApiDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/v1/groups/{groupId}",
        produces = { "application/json" }
    )
    ResponseEntity<GroupApiDto> findGroup(
        @Parameter(name = "groupId", description = "Group id", required = true) @PathVariable("groupId") UUID groupId
    );


    /**
     * GET /v1/groups/search/{name} : Get a group by name.
     *
     * @param name Group name (required)
     * @return OK (status code 200)
     *         or error (status code 400)
     */
    @Operation(
        operationId = "findGroupByName",
        summary = "Get a group by name.",
        tags = { "groups-v1" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GroupApiDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorApiDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/v1/groups/search/{name}",
        produces = { "application/json" }
    )
    ResponseEntity<GroupApiDto> findGroupByName(
        @Parameter(name = "name", description = "Group name", required = true) @PathVariable("name") String name
    );


    /**
     * GET /v1/groups : Get a groups. Results are pageable with seek method.
     *
     * @return OK (status code 200)
     *         or error (status code 400)
     */
    @Operation(
        operationId = "findGroups",
        summary = "Get a groups. Results are pageable with seek method.",
        tags = { "groups-v1" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GroupApiDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorApiDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/v1/groups",
        produces = { "application/json" }
    )
    ResponseEntity<List<GroupApiDto>> findGroups(
        
    );


    /**
     * DELETE /v1/groups/{groupId} : Delete a group.
     *
     * @param groupId Group id (required)
     * @return OK (status code 200)
     *         or error (status code 400)
     */
    @Operation(
        operationId = "removeGroup",
        summary = "Delete a group.",
        tags = { "groups-v1" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GroupApiDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorApiDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/v1/groups/{groupId}",
        produces = { "application/json" }
    )
    ResponseEntity<GroupApiDto> removeGroup(
        @Parameter(name = "groupId", description = "Group id", required = true) @PathVariable("groupId") UUID groupId
    );


    /**
     * PUT /v1/groups/{groupId} : Update a group.
     *
     * @param groupId Group id (required)
     * @param groupApiDto  (required)
     * @return OK (status code 200)
     *         or error (status code 400)
     */
    @Operation(
        operationId = "updateGroup",
        summary = "Update a group.",
        tags = { "groups-v1" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GroupApiDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorApiDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/v1/groups/{groupId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<GroupApiDto> updateGroup(
        @Parameter(name = "groupId", description = "Group id", required = true) @PathVariable("groupId") UUID groupId,
        @Parameter(name = "GroupApiDto", description = "", required = true) @Valid @RequestBody GroupApiDto groupApiDto
    );

}